+++ NEWS.md
@@ -1,6 +1,7 @@
 
 ## Version 1.1.0 (development)
 
+* `google_place_details()` implemented to retrieve more details about a specific place
 * `google_places()` implemented to retrieve information from Google Places API
 * `traffic_model` arguments correctly defined - closes [#16](https://github.com/SymbolixAU/googleway/issues/16)
 * Waypoint list elements must be named ('stop' or 'via') - closes [#12](https://github.com/SymbolixAU/googleway/issues/12)